feat: apply to `TagsPage`
