+++ extensions/tags/js/src/forum/addTagList.js
@@ -1,17 +1,17 @@
+import app from 'flarum/forum/app';
 import { extend } from 'flarum/common/extend';
-import IndexPage from 'flarum/forum/components/IndexPage';
+import IndexSidebar from 'flarum/forum/components/IndexSidebar';
 import Separator from 'flarum/common/components/Separator';
 import LinkButton from 'flarum/common/components/LinkButton';
 
 import TagLinkButton from './components/TagLinkButton';
 import TagsPage from './components/TagsPage';
-import app from 'flarum/forum/app';
 import sortTags from '../common/utils/sortTags';
 
 export default function addTagList() {
   // Add a link to the tags page, as well as a list of all the tags,
   // to the index page's sidebar.
-  extend(IndexPage.prototype, 'navItems', function (items) {
+  extend(IndexSidebar.prototype, 'navItems', function (items) {
     items.add(
       'tags',
       <LinkButton icon="fas fa-th-large" href={app.route('tags')}>
@@ -26,7 +26,7 @@ export default function addTagList() {
 
     const params = app.search.stickyParams();
     const tags = app.store.all('tags');
-    const currentTag = this.currentTag();
+    const currentTag = app.currentTag();
 
     const addTag = (tag) => {
       let active = currentTag === tag;
