+++ framework/core/js/src/forum/components/IndexPage.tsx
@@ -13,6 +13,7 @@ import extractText from '../../common/utils/extractText';
 import type Mithril from 'mithril';
 import type Discussion from '../../common/models/Discussion';
 import PageStructure from './PageStructure';
+import IndexSidebar from './IndexSidebar';
 
 export interface IIndexPageAttrs extends IPageAttrs {}
 
@@ -133,71 +134,8 @@ export default class IndexPage<CustomAttrs extends IIndexPageAttrs = IIndexPageA
     return <WelcomeHero />;
   }
 
-  /**
-   * Build an item list for the sidebar of the index page. By default this is a
-   * "New Discussion" button, and then a DropdownSelect component containing a
-   * list of navigation items.
-   */
-  sidebarItems() {
-    const items = new ItemList<Mithril.Children>();
-    const canStartDiscussion = app.forum.attribute('canStartDiscussion') || !app.session.user;
-
-    items.add(
-      'newDiscussion',
-      <Button
-        icon="fas fa-edit"
-        className="Button Button--primary IndexPage-newDiscussion"
-        itemClassName="App-primaryControl"
-        onclick={() => {
-          // If the user is not logged in, the promise rejects, and a login modal shows up.
-          // Since that's already handled, we dont need to show an error message in the console.
-          return this.newDiscussionAction().catch(() => {});
-        }}
-        disabled={!canStartDiscussion}
-      >
-        {app.translator.trans(`core.forum.index.${canStartDiscussion ? 'start_discussion_button' : 'cannot_start_discussion_button'}`)}
-      </Button>
-    );
-
-    items.add(
-      'nav',
-      <SelectDropdown
-        buttonClassName="Button"
-        className="App-titleControl"
-        accessibleToggleLabel={app.translator.trans('core.forum.index.toggle_sidenav_dropdown_accessible_label')}
-      >
-        {this.navItems().toArray()}
-      </SelectDropdown>
-    );
-
-    return items;
-  }
-
   sidebar() {
-    return (
-      <nav className="IndexPage-nav sideNav">
-        <ul>{listItems(this.sidebarItems().toArray())}</ul>
-      </nav>
-    );
-  }
-
-  /**
-   * Build an item list for the navigation in the sidebar of the index page. By
-   * default this is just the 'All Discussions' link.
-   */
-  navItems() {
-    const items = new ItemList<Mithril.Children>();
-    const params = app.search.stickyParams();
-
-    items.add(
-      'allDiscussions',
-      <LinkButton href={app.route('index', params)} icon="far fa-comments">
-        {app.translator.trans('core.forum.index.all_discussions_link')}
-      </LinkButton>,
-      100
-    );
-
-    return items;
+    return <IndexSidebar />;
   }
 
   /**
@@ -273,23 +211,6 @@ export default class IndexPage<CustomAttrs extends IIndexPageAttrs = IIndexPageA
     }
 
     return items;
-  }
-
-  /**
-   * Open the composer for a new discussion or prompt the user to login.
-   */
-  newDiscussionAction(): Promise<unknown> {
-    return new Promise((resolve, reject) => {
-      if (app.session.user) {
-        app.composer.load(() => import('./DiscussionComposer'), { user: app.session.user }).then(() => app.composer.show());
-
-        return resolve(app.composer);
-      } else {
-        app.modal.show(() => import('./LogInModal'));
-
-        return reject();
-      }
-    });
   }
 
   /**
