+++ framework/core/less/common/sideNav.less
@@ -58,14 +58,11 @@
   }
   .sideNav {
     flex-shrink: 0;
-    margin-right: 50px;
 
     &, > ul {
       width: 190px;
     }
     > ul {
-      margin-top: 30px;
-
       &.affix {
         top: var(--header-height);
       }
