+++ framework/core/less/forum/DiscussionPage.less
@@ -35,6 +35,10 @@
       &-content {
         margin-top: 10px;
       }
+
+      &-sidebar {
+        margin-top: 0;
+      }
     }
   }
