+++ framework/core/less/forum/PageStructure.less
@@ -14,13 +14,16 @@
     }
 
     &-content {
-      margin-top: 30px;
       width: var(--content-width);
     }
 
     &-sidebar {
       width: var(--sidebar-width);
       flex-shrink: 0;
+    }
+
+    &-content, &-sidebar {
+      margin-top: 30px;
     }
   }
